+++ lua/neotest-haskell/init.lua
@@ -1,11 +1,11 @@
-local lib = require('neotest.lib')
 local base = require('neotest-haskell.base')
+local hspec = require('neotest-haskell.hspec')
 local cabal = require("neotest-haskell.cabal")
 local stack = require("neotest-haskell.stack")
+local lib = require('neotest.lib')
 local logger = require("neotest.logging")
 
 
----@type neotest.Adapter
 local HaskellNeotestAdapter = { name = 'neotest-haskell' }
 
 
@@ -23,25 +23,22 @@ function HaskellNeotestAdapter.filter_dir(...)
 end
 
 ---@async
----@return neotest.Tree | nil
 function HaskellNeotestAdapter.discover_positions(path)
-  local pos = base.parse_positions(path)
+  local pos = hspec.parse_positions(path)
   logger.debug("Found positions: " .. vim.inspect(pos))
   return pos
 end
 
 
 ---@async
----@param args neotest.RunArgs
----@return neotest.RunSpec
 function HaskellNeotestAdapter.build_spec(args)
-  local tree = args.tree
+  local tree = args and args.tree
   if not tree then
-    return
+    return {}
   end
   local pos = args.tree:data()
   if pos.type ~= "test" then
-    return
+    return {}
   end
 
   local mkCommand = function(command)
@@ -55,14 +52,13 @@ function HaskellNeotestAdapter.build_spec(args)
     }
   end
 
-  local hspec_match = base.get_hspec_match(pos)
   local package_root = base.match_package_root_pattern(pos.path)
   local project_root = base.match_project_root_pattern(pos.path)
 
   if lib.files.exists(project_root .. '/cabal.project') then
-    return mkCommand(cabal.build_command(package_root, hspec_match))
+    return mkCommand(cabal.build_command(package_root, pos))
   elseif lib.files.exists(project_root .. '/package.yaml') then
-    return mkCommand(stack.build_command(project_root, package_root, hspec_match))
+    return mkCommand(stack.build_command(project_root, package_root, pos))
   end
 
   logger.error('Project is neither a Cabal nor a Stack project.')
@@ -70,9 +66,6 @@ end
 
 
 ---@async
----@param spec neotest.RunSpec
----@param result neotest.StrategyResult
----@return neotest.Result[]
 function HaskellNeotestAdapter.results(spec, result)
   local pos_id = spec.context.pos_id
   if result.code == 0 then